firebase.js lista
