--- src/model.js
@@ -1,13 +1,0 @@
-
-var firebaseConfig = {
-  apiKey: "AIzaSyAphkTjnCyuMEe9J2BlkLSnRf11LDrRKq8",
-  authDomain: "jaloredsocial.firebaseapp.com",
-  projectId: "jaloredsocial",
-  storageBucket: "jaloredsocial.appspot.com",
-  messagingSenderId: "438968128013",
-  appId: "1:438968128013:web:9d1b47242a6f58c825bb44",
-  measurementId: "G-8FRZGM62BF"
-};
-// Initialize Firebase
-firebase.initializeApp(firebaseConfig);
-firebase.analytics();
